third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-content.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style>
#target1::before {
  content: 'test';
}
#target1::after {
  content: normal;
}
#target2 {
  content: 'test';
}
#target2::before {
  content: normal;
}
#target2::after {
  content: 'test';
}
#target3 {
  content: none;
}
#target3::before {
  content: none;
}
#target3::after {
  content: 'test';
}
</style>

<div id="target1"></div>
<div id="target2"></div>
<div id="target3"></div>

<script>
test(function() {
  assert_equals(getComputedStyle(target1).content, 'normal');
  assert_equals(getComputedStyle(target1, '::before').content, '"test"');
  assert_equals(getComputedStyle(target1, '::after').content, 'none');
  assert_equals(getComputedStyle(target2).content, 'normal');
  assert_equals(getComputedStyle(target2, '::before').content, 'none');
  assert_equals(getComputedStyle(target2, '::after').content, '"test"');
  assert_equals(getComputedStyle(target3).content, 'normal');
  assert_equals(getComputedStyle(target3, '::before').content, 'none');
  assert_equals(getComputedStyle(target3, '::after').content, '"test"');
}, 'Test that content computes to "normal" for elements (not pseudo elements) and to "none" if the specified value of pseudo elements is "normal"');
</script>
